refactor(app): move app creation to factory method
app was untestable, so app-factory.js was created to ensure it can be tested.

closes #141
